+++ nym-connect/src-tauri/src/operations/connection/connect.rs
@@ -1,6 +1,8 @@
-use crate::error::BackendError;
-use crate::models::ConnectResult;
-use crate::State;
+use crate::{
+    error::{BackendError, Result},
+    models::ConnectResult,
+    tasks, State,
+};
 use std::sync::Arc;
 use tokio::sync::RwLock;
 
@@ -8,21 +10,23 @@ use tokio::sync::RwLock;
 pub async fn start_connecting(
     state: tauri::State<'_, Arc<RwLock<State>>>,
     window: tauri::Window<tauri::Wry>,
-) -> Result<ConnectResult, BackendError> {
-    let mut guard = state.write().await;
+) -> Result<ConnectResult> {
+    let status_receiver = {
+        let mut state_w = state.write().await;
+        state_w.start_connecting(&window).await?
+    };
 
-    guard.start_connecting(&window).await;
+    // Setup task for checking status
+    let state = state.inner().clone();
+    tasks::start_disconnect_listener(state, window, status_receiver);
 
     Ok(ConnectResult {
-        // WIP(JON): fixme
-        address: "Test".to_string(),
+        address: "PLACEHOLDER".to_string(),
     })
 }
 
 #[tauri::command]
-pub async fn get_service_provider(
-    state: tauri::State<'_, Arc<RwLock<State>>>,
-) -> Result<String, BackendError> {
+pub async fn get_service_provider(state: tauri::State<'_, Arc<RwLock<State>>>) -> Result<String> {
     let guard = state.read().await;
     guard
         .get_service_provider()
@@ -34,16 +38,15 @@ pub async fn get_service_provider(
 pub async fn set_service_provider(
     service_provider: String,
     state: tauri::State<'_, Arc<RwLock<State>>>,
-) -> Result<(), BackendError> {
+) -> Result<()> {
+    log::trace!("Setting service_provider: {service_provider}");
     let mut guard = state.write().await;
     guard.set_service_provider(service_provider);
     Ok(())
 }
 
 #[tauri::command]
-pub async fn get_gateway(
-    state: tauri::State<'_, Arc<RwLock<State>>>,
-) -> Result<String, BackendError> {
+pub async fn get_gateway(state: tauri::State<'_, Arc<RwLock<State>>>) -> Result<String> {
     let guard = state.read().await;
     guard
         .get_gateway()
@@ -55,7 +58,8 @@ pub async fn get_gateway(
 pub async fn set_gateway(
     gateway: String,
     state: tauri::State<'_, Arc<RwLock<State>>>,
-) -> Result<(), BackendError> {
+) -> Result<()> {
+    log::trace!("Setting gateway: {gateway}");
     let mut guard = state.write().await;
     guard.set_gateway(gateway);
     Ok(())