+++ nym-connect/src/context/main.tsx
@@ -4,6 +4,7 @@ import { invoke } from '@tauri-apps/api/tauri';
 import { listen, UnlistenFn } from '@tauri-apps/api/event';
 import { ConnectionStatusKind } from '../types';
 import { ConnectionStatsItem } from '../components/ConnectionStats';
+import { ServiceProvider, Services } from '../types/directory';
 
 const TAURI_EVENT_STATUS_CHANGED = 'app:connection-status-changed';
 
@@ -14,11 +15,14 @@ type TClientContext = {
   connectionStatus: ConnectionStatusKind;
   connectionStats?: ConnectionStatsItem[];
   connectedSince?: DateTime;
+  services?: Services;
+  serviceProvider?: ServiceProvider;
 
   setMode: (mode: ModeType) => void;
   setConnectionStatus: (connectionStatus: ConnectionStatusKind) => void;
   setConnectionStats: (connectionStats: ConnectionStatsItem[] | undefined) => void;
   setConnectedSince: (connectedSince: DateTime | undefined) => void;
+  setServiceProvider: (serviceProvider: ServiceProvider) => void;
 
   startConnecting: () => Promise<void>;
   startDisconnecting: () => Promise<void>;
@@ -31,6 +35,14 @@ export const ClientContextProvider = ({ children }: { children: React.ReactNode 
   const [connectionStatus, setConnectionStatus] = useState<ConnectionStatusKind>(ConnectionStatusKind.disconnected);
   const [connectionStats, setConnectionStats] = useState<ConnectionStatsItem[]>();
   const [connectedSince, setConnectedSince] = useState<DateTime>();
+  const [services, setServices] = React.useState<Services>();
+  const [serviceProvider, setRawServiceProvider] = React.useState<ServiceProvider>();
+
+  useEffect(() => {
+    invoke('get_services').then((result) => {
+      setServices(result as Services);
+    });
+  }, []);
 
   useEffect(() => {
     let unlisten: UnlistenFn | undefined;
@@ -59,6 +71,12 @@ export const ClientContextProvider = ({ children }: { children: React.ReactNode 
     await invoke('start_disconnecting');
   }, []);
 
+  const setServiceProvider = useCallback(async (newServiceProvider: ServiceProvider) => {
+    await invoke('set_gateway', { gateway: newServiceProvider.gateway });
+    await invoke('set_service_provider', { serviceProvider: newServiceProvider.address });
+    setRawServiceProvider(newServiceProvider);
+  }, []);
+
   return (
     <ClientContext.Provider
       value={{
@@ -72,6 +90,9 @@ export const ClientContextProvider = ({ children }: { children: React.ReactNode 
         setConnectedSince,
         startConnecting,
         startDisconnecting,
+        services,
+        serviceProvider,
+        setServiceProvider,
       }}
     >
       {children}