+++ nym-connect/src/stories/AppFlow.stories.tsx
@@ -7,6 +7,7 @@ import { useClientContext } from '../context/main';
 import { ConnectionStatusKind } from '../types';
 import { DefaultLayout } from '../layouts/DefaultLayout';
 import { ConnectedLayout } from '../layouts/ConnectedLayout';
+import { Services } from '../types/directory';
 
 export default {
   title: 'App/Flow',
@@ -16,6 +17,19 @@ export default {
 export const Mock: ComponentStory<typeof AppWindowFrame> = () => {
   const context = useClientContext();
   const [busy, setBusy] = React.useState<boolean>();
+  const services: Services = [
+    {
+      id: 'keybase',
+      description: 'Keybase',
+      items: [
+        {
+          id: 'nym-keybase',
+          description: 'Nym Keybase Service Provider',
+          address: '1234.5678',
+        },
+      ],
+    },
+  ];
   const handleConnectClick = React.useCallback(() => {
     const oldStatus = context.connectionStatus;
     if (oldStatus === ConnectionStatusKind.connected || oldStatus === ConnectionStatusKind.disconnected) {
@@ -53,7 +67,12 @@ export const Mock: ComponentStory<typeof AppWindowFrame> = () => {
   ) {
     return (
       <Box p={4} sx={{ background: 'white' }}>
-        <DefaultLayout status={context.connectionStatus} busy={busy} onConnectClick={handleConnectClick} />
+        <DefaultLayout
+          status={context.connectionStatus}
+          busy={busy}
+          onConnectClick={handleConnectClick}
+          services={services}
+        />
       </Box>
     );
   }
@@ -67,6 +86,7 @@ export const Mock: ComponentStory<typeof AppWindowFrame> = () => {
         ipAddress="127.0.0.1"
         port={1080}
         connectedSince={context.connectedSince}
+        serviceProvider={services[0].items[0]}
         stats={[
           {
             label: 'in:',